Split vim lua configs
