+++ home-manager/modules/vim/lua/config/bufferline.lua
@@ -1,0 +1,48 @@
+require'bufferline'.setup{
+	options = {
+		numbers =  "ordinal",
+		number_style = "normal",
+		mappings = true, --  <leader>1-9 mappings to navigate tabs
+		buffer_close_icon= '',
+		modified_icon = '●',
+		close_icon = '',
+		left_trunc_marker = '',
+		right_trunc_marker = '',
+		max_name_length = 18,
+		max_prefix_length = 15, -- prefix used when a buffer is deduplicated
+		tab_size = 18,
+		diagnostics = "nvim_lsp",
+		diagnostics_indicator = function(count, level, diagnostics_dict)
+			return "("..count..")"
+		end,
+		-- NOTE: this will be called a lot so don't do any heavy processing here
+		-- custom_filter = function(buf_number)
+		--   -- filter out filetypes you don't want to see
+		--   if vim.bo[buf_number].filetype ~= "<i-dont-want-to-see-this>" then
+		--     return true
+		--   end
+		--   -- filter out by buffer name
+		--   if vim.fn.bufname(buf_number) ~= "<buffer-name-I-dont-want>" then
+		--     return true
+		--   end
+		--   -- filter out based on arbitrary rules
+		--   -- e.g. filter out vim wiki buffer from tabline in your work repo
+		--   if vim.fn.getcwd() == "<work-repo>" and vim.bo[buf_number].filetype ~= "wiki" then
+		--     return true
+		--   end
+		-- end,
+		show_buffer_close_icons = true,
+		show_close_icon = true,
+		show_tab_indicators = true,
+		persist_buffer_sort = true, -- whether or not custom sorted buffers should persist
+		-- can also be a table containing 2 custom separators
+		-- [focused and unfocused]. eg: { '|', '|' }
+		separator_style = "slant", --| "thick" | "thin" | { 'any', 'any' },
+		enforce_regular_tabs = false , --| true,
+		always_show_bufferline = true, --, | false,
+		-- sort_by =  function(buffer_a, buffer_b)
+		--   -- add custom logic
+		--   return buffer_a.modified > buffer_b.modified
+		-- end
+	},
+}
