+++ home-manager/modules/vim/lua/config/devicons.lua
@@ -1,0 +1,5 @@
+require'nvim-web-devicons'.setup {
+	-- globally enable default icons (default to false)
+	-- will get overriden by `get_icons` option
+	default = true;
+}
