+++ home-manager/modules/vim/lua/config/lsp.lua
@@ -1,0 +1,17 @@
+require'lspconfig'.pyright.setup{}
+require'lspconfig'.gopls.setup{}
+require'lspconfig'.terraformls.setup{}
+require'lspconfig'.bashls.setup{}
+require'lspconfig'.yamlls.setup{}
+
+require'lspconfig'.jsonls.setup {
+
+	cmd = { "json-languageserver", "--stdio" },
+	commands = {
+		Format = {
+			function()
+				vim.lsp.buf.range_formatting({},{0,0},{vim.fn.line("$"),0})
+			end
+		}
+	}
+}
